5260 Extract UploadDockerImages from UploadDockerImagesToRepository
